Sort countries alphabeitcally in the AP settings
This patch changes the country list for the wireless AP settings so they are
always sorted alphabetically, taking into account i18n. The sorting is done
using locale.strcoll as the comparison function so reasonably correct
sorting is expected for various languages.

Because we are wrapping the sorting code in bottle_utils.lazy.Lazy, the
null-value item is set using empty_value kwarg, which requires
Outernet-Project/librarian#340 to be merged.

Signed-off-by: Branko Vukelic <branko@outernet.is>